fix: move tests from modules:data-streams to x-pack:logsdb
