+++ x-pack/plugin/logsdb/src/javaRestTest/java/org/elasticsearch/xpack/logsdb/qa/matchers/ArrayEqualMatcher.java
@@ -1,22 +1,17 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
-package org.elasticsearch.datastreams.logsdb.qa.matchers;
+package org.elasticsearch.xpack.logsdb.qa.matchers;
 
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.xcontent.XContentBuilder;
 
 import java.util.Arrays;
 import java.util.List;
-
-import static org.elasticsearch.datastreams.logsdb.qa.matchers.Messages.formatErrorMessage;
-import static org.elasticsearch.datastreams.logsdb.qa.matchers.Messages.prettyPrintArrays;
 
 class ArrayEqualMatcher extends GenericEqualsMatcher<Object[]> {
     ArrayEqualMatcher(
@@ -39,12 +34,12 @@ class ArrayEqualMatcher extends GenericEqualsMatcher<Object[]> {
     private MatchResult matchArraysEqual(final Object[] actualArray, final Object[] expectedArray, boolean ignoreSorting) {
         if (actualArray.length != expectedArray.length) {
             return MatchResult.noMatch(
-                formatErrorMessage(
+                Messages.formatErrorMessage(
                     actualMappings,
                     actualSettings,
                     expectedMappings,
                     expectedSettings,
-                    "Array lengths do no match, " + prettyPrintArrays(actualArray, expectedArray)
+                    "Array lengths do no match, " + Messages.prettyPrintArrays(actualArray, expectedArray)
                 )
             );
         }
@@ -52,24 +47,24 @@ class ArrayEqualMatcher extends GenericEqualsMatcher<Object[]> {
             return matchArraysEqualIgnoringSorting(actualArray, expectedArray)
                 ? MatchResult.match()
                 : MatchResult.noMatch(
-                    formatErrorMessage(
+                    Messages.formatErrorMessage(
                         actualMappings,
                         actualSettings,
                         expectedMappings,
                         expectedSettings,
-                        "Arrays do not match when ignoring sort order, " + prettyPrintArrays(actualArray, expectedArray)
+                        "Arrays do not match when ignoring sort order, " + Messages.prettyPrintArrays(actualArray, expectedArray)
                     )
                 );
         } else {
             return matchArraysEqualExact(actualArray, expectedArray)
                 ? MatchResult.match()
                 : MatchResult.noMatch(
-                    formatErrorMessage(
+                    Messages.formatErrorMessage(
                         actualMappings,
                         actualSettings,
                         expectedMappings,
                         expectedSettings,
-                        "Arrays do not match exactly, " + prettyPrintArrays(actualArray, expectedArray)
+                        "Arrays do not match exactly, " + Messages.prettyPrintArrays(actualArray, expectedArray)
                     )
                 );
         }
