+++ x-pack/plugin/logsdb/src/javaRestTest/java/org/elasticsearch/xpack/logsdb/qa/matchers/GenericEqualsMatcher.java
@@ -1,20 +1,16 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
-package org.elasticsearch.datastreams.logsdb.qa.matchers;
+package org.elasticsearch.xpack.logsdb.qa.matchers;
 
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.xcontent.XContentBuilder;
 
 import java.util.List;
-
-import static org.elasticsearch.datastreams.logsdb.qa.matchers.Messages.formatErrorMessage;
 
 public class GenericEqualsMatcher<T> extends Matcher {
     protected final XContentBuilder actualMappings;
@@ -48,7 +44,7 @@ public class GenericEqualsMatcher<T> extends Matcher {
             if (expected == null) {
 
                 return MatchResult.noMatch(
-                    formatErrorMessage(
+                    Messages.formatErrorMessage(
                         actualMappings,
                         actualSettings,
                         expectedMappings,
@@ -58,17 +54,29 @@ public class GenericEqualsMatcher<T> extends Matcher {
                 );
             }
             return MatchResult.noMatch(
-                formatErrorMessage(actualMappings, actualSettings, expectedMappings, expectedSettings, "Expected is null but actual is not")
+                Messages.formatErrorMessage(
+                    actualMappings,
+                    actualSettings,
+                    expectedMappings,
+                    expectedSettings,
+                    "Expected is null but actual is not"
+                )
             );
         }
         if (expected == null) {
             return MatchResult.noMatch(
-                formatErrorMessage(actualMappings, actualSettings, expectedMappings, expectedSettings, "Actual is null but expected is not")
+                Messages.formatErrorMessage(
+                    actualMappings,
+                    actualSettings,
+                    expectedMappings,
+                    expectedSettings,
+                    "Actual is null but expected is not"
+                )
             );
         }
         if (actual.getClass().equals(expected.getClass()) == false) {
             return MatchResult.noMatch(
-                formatErrorMessage(
+                Messages.formatErrorMessage(
                     actualMappings,
                     actualSettings,
                     expectedMappings,
