+++ x-pack/plugin/logsdb/src/javaRestTest/java/org/elasticsearch/xpack/logsdb/qa/matchers/source/FieldSpecificMatcher.java
@@ -1,18 +1,17 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
-package org.elasticsearch.datastreams.logsdb.qa.matchers.source;
+package org.elasticsearch.xpack.logsdb.qa.matchers.source;
 
 import org.apache.lucene.sandbox.document.HalfFloatPoint;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.datastreams.logsdb.qa.matchers.MatchResult;
 import org.elasticsearch.xcontent.XContentBuilder;
+import org.elasticsearch.xpack.logsdb.qa.matchers.MatchResult;
+import org.elasticsearch.xpack.logsdb.qa.matchers.Messages;
 
 import java.math.BigInteger;
 import java.util.List;
@@ -21,9 +20,6 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-
-import static org.elasticsearch.datastreams.logsdb.qa.matchers.Messages.formatErrorMessage;
-import static org.elasticsearch.datastreams.logsdb.qa.matchers.Messages.prettyPrintCollections;
 
 interface FieldSpecificMatcher {
     MatchResult match(List<Object> actual, List<Object> expected, Map<String, Object> actualMapping, Map<String, Object> expectedMapping);
@@ -59,13 +55,13 @@ interface FieldSpecificMatcher {
             return actualHalfFloatBytes.equals(expectedHalfFloatBytes)
                 ? MatchResult.match()
                 : MatchResult.noMatch(
-                    formatErrorMessage(
+                    Messages.formatErrorMessage(
                         actualMappings,
                         actualSettings,
                         expectedMappings,
                         expectedSettings,
                         "Values of type [half_float] don't match after normalization, normalized "
-                            + prettyPrintCollections(actualHalfFloatBytes, expectedHalfFloatBytes)
+                            + Messages.prettyPrintCollections(actualHalfFloatBytes, expectedHalfFloatBytes)
                     )
                 );
         }
@@ -123,13 +119,13 @@ interface FieldSpecificMatcher {
             return actualNormalized.equals(expectedNormalized)
                 ? MatchResult.match()
                 : MatchResult.noMatch(
-                    formatErrorMessage(
+                    Messages.formatErrorMessage(
                         actualMappings,
                         actualSettings,
                         expectedMappings,
                         expectedSettings,
                         "Values of type [scaled_float] don't match after normalization, normalized "
-                            + prettyPrintCollections(actualNormalized, expectedNormalized)
+                            + Messages.prettyPrintCollections(actualNormalized, expectedNormalized)
                     )
                 );
         }
@@ -194,13 +190,13 @@ interface FieldSpecificMatcher {
             return actualNormalized.equals(expectedNormalized)
                 ? MatchResult.match()
                 : MatchResult.noMatch(
-                    formatErrorMessage(
+                    Messages.formatErrorMessage(
                         actualMappings,
                         actualSettings,
                         expectedMappings,
                         expectedSettings,
                         "Values of type [unsigned_long] don't match after normalization, normalized "
-                            + prettyPrintCollections(actualNormalized, expectedNormalized)
+                            + Messages.prettyPrintCollections(actualNormalized, expectedNormalized)
                     )
                 );
         }
